Added new fcns PRM now builds without checking for obstacle Only A* and postprocesspath2 check for obstacle This way we can reuse PRM depending on where obstacale is
Next step: PRM is deleted everytime we exit findPlan
Find a way to store PRM for reusing
Check if A* works for new start state?
Is it already in graph or do we need to manually insert it into PRM?

Questions to answer:
Is A* fast enough to be used in real time?
How to store PRM for reusing?
Should we change heuristic value of obstacle nodes in PRM automatically?
Would go into bounding box calculations
I think doing constant isValidState Checks  might be slower

D* or backward A* for replanning?